+++ vgo-core/src/main/kotlin/com/jzbrooks/vgo/core/optimization/BottomUpOptimization.kt
@@ -1,9 +1,0 @@
-package com.jzbrooks.vgo.core.optimization
-
-import com.jzbrooks.vgo.core.graphic.ElementVisitor
-
-@Deprecated(
-    "Has been relocated to the transformation package",
-    replaceWith = ReplaceWith("com.jzbrooks.vgo.core.transformation.BottomUpTransformation"),
-)
-interface BottomUpOptimization : ElementVisitor