+++ vgo-gradle/src/main/kotlin/com/jzbrooks/vgo/plugin/VgoPlugin.kt
@@ -9,14 +9,6 @@ class VgoPlugin : Plugin<Project> {
     override fun apply(target: Project) {
         target.extensions.create<VgoPluginExtension>("vgo")
 
-        target.tasks.register<ShrinkVectorArtwork>("shrinkVectorArtwork") {
-            group = "resource"
-            description = "Shrinks vector graphic files. Deprecated. Use shrinkVectorGraphic instead."
-            doFirst {
-                logger.warn("This task is deprecated. Use shrinkVectorGraphic instead.")
-            }
-        }
-
         target.tasks.register<ShrinkVectorGraphic>("shrinkVectorGraphic") {
             group = "resource"
             description = "Shrinks vector graphic files."